Requested Changes - renamed file - code style - added types
